% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agb_example}
\alias{agb_example}
\title{Path to directory containing example data.}
\usage{
agb_example(path)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/.}
}
\value{
Path to directory containing example data.
}
\description{
Path to directory containing example data.
}
\examples{
agb_example("data")
}
