% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes-package.R
\docType{package}
\name{AGBfluxes-package}
\alias{AGBfluxes}
\alias{AGBfluxes-package}
\title{AGBfluxes: Compute Flux of Biomass at ForestGEO Sites}
\description{
Compute flux of biomass at ForestGEO sites.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/AGBfluxes}
  \item Report bugs at \url{https://github.com/forestgeo/AGBfluxes/issues}
}

}
\author{
\strong{Maintainer}: Ervan Rutishauser \email{er.rutishauser@gmail.com} [contractor]

Authors:
\itemize{
  \item Mauro Lepore \email{leporem@si.edu}
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder]
}

}
\keyword{internal}
