% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{assign_status}
\alias{assign_status}
\title{Assign status.}
\usage{
assign_status(DT)
}
\arguments{
\item{DF}{A data.table.}
}
\value{
Update the column 'status1' with consistent information.
}
\description{
Assign status alive ("A"), alive with POM changed ("AC"), dead ("D"),
recruited ("R") or resprout ("Rsp") to trees, and check for consistency over
time (i.e. avoid resurrection).
}
