% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{assignWD}
\alias{assignWD}
\title{Assign wood density.}
\usage{
assignWD(DAT, WD = NULL)
}
\arguments{
\item{df}{A data.table.}

\item{site}{Provide the full name of your site (in lower case) i.e. 'barro
colorado island'.}

\item{wsgdata}{A list of tree species (mnenomic) and corresponding wood
density.}

\item{denscol}{The variable to be return ("wsg" by default). No other option
is implemented.}
}
\value{
A data.table (data.frame) with all relevant variables.
}
\description{
Assign wood density using CTFS wood density data base (WSG).
}
