% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci.spptable}
\alias{bci.spptable}
\title{Species table from from Barro Colorado Island (BCI).}
\format{An object of class \code{data.frame} with 1053 rows and 10 columns.}
\usage{
bci.spptable
}
\description{
Species table from from Barro Colorado Island (BCI).
}
\examples{
str(bci.spptable)
}
\keyword{datasets}
