% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_stem_year}
\alias{bci_stem_year}
\alias{bci_stem_1995}
\alias{bci_stem_2000}
\alias{bci_stem_2005}
\title{Example stem-tables from Barro Colorado Island (BCI).}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 0 rows and 17 columns.}
\usage{
bci_stem_1995

bci_stem_2000

bci_stem_2005
}
\description{
Example stem-tables from Barro Colorado Island (BCI).
}
\examples{
str(bci_stem_1995)
str(bci_stem_2000)
str(bci_stem_2000)
}
\keyword{datasets}
