% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{computeAGB}
\alias{computeAGB}
\title{Biomass computation.}
\usage{
computeAGB(df, use_palm_allometry, DBH = NULL, WD = NULL, H = NULL)
}
\arguments{
\item{df}{The data.frame on which AGB shall be computed.}

\item{use_palm_allometry}{\code{TRUE} or \code{FALSE}, if \code{TRUE}  biomass of palm trees
is computed through a specific allometric model (Goodman et al. 2013). Only
valid for South America.}

\item{DBH}{Optional, specify the variable to be used (e.g. "dbh" or
"dbh_corrected").}

\item{WD}{Optional, provide an external data.frame of wood densities by
species.}

\item{H}{Optional, specify the column with height measurments.}
}
\value{
A data.table (data.frame) with all relevant variables.
}
\description{
Allocate wood density and compute above-ground biomass using the updated
model of Chave et al. (2014), given in Rejou-Mechain et al. (2017). Palm
trees (palm=T) are computed using a different allometric model (Goodman et
al. 2013).
}
