% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{consolidate_data}
\alias{consolidate_data}
\title{This function checks for each stem that (i) its status (i.e. alive or dead) is consistent across all censuses (e.g. can't be dead and alive at next census), (ii) assign a mean census date (when missing) and (iii) interpolate DBHs when missing (fill_missing=T). Additionnaly, unecessary information (i.e. replication of dead trees/recruits) is discarded.}
\usage{
consolidate_data(df, taper_correction, fill_missing, stem)
}
\arguments{
\item{taper_correction}{\code{TRUE} or \code{FALSE}, are you willing to apply Cushman
et al (2014) taper correction?}

\item{fill_missing}{\code{TRUE} or \code{FALSE}, are you willing to extrapolate missing
DBH from surrounding DBHs?}

\item{stem}{Is the function applied over stem (stem=TRUE) or tree
(stem=FALSE) data?}
}
\value{
A data.table (data.frame) with 3 new variables: status1 (corrected status),dbh2 (corrected dbh) and hom2 (corrected height of measure). A unique stem id ("id") is created if stem = T.
}
\description{
This function checks for each stem that (i) its status (i.e. alive or dead) is consistent across all censuses (e.g. can't be dead and alive at next census), (ii) assign a mean census date (when missing) and (iii) interpolate DBHs when missing (fill_missing=T). Additionnaly, unecessary information (i.e. replication of dead trees/recruits) is discarded.
}
\examples{
data(df)
# Correct for change in POM and missing value
df <- consolidate_data(df, taper_correction=T, fill_missing=T, stem=T)
df[treeID==240] # show the result with missing value
# Correct for missing value only
df <- consolidate_data(df, taper_correction=F, fill_missing=T, stem=T)
df[treeID==240] # show the result with missing value
# No correction
df <- consolidate_data(df, taper_correction=F, fill_missing=F, stem=T)
df[treeID==240] # show the result with missing value
}
