% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R, R/data_preparation.R
\name{data_preparation}
\alias{data_preparation}
\title{CTFS-formated data preparation}
\usage{
data_preparation(path, stem, site, dbh_units = "mm", WD = NULL,
  taper_correction = TRUE, fill_missing = TRUE,
  use_palm_allometry = TRUE, flag_stranglers = TRUE,
  dbh_stranglers = 500, maxrel = 0.2, write_errors_to = NULL,
  graph_problems_to = NULL, exclude_interval = NULL)

data_preparation(path, stem, site, dbh_units = "mm", WD = NULL,
  taper_correction = TRUE, fill_missing = TRUE,
  use_palm_allometry = TRUE, flag_stranglers = TRUE,
  dbh_stranglers = 500, maxrel = 0.2, write_errors_to = NULL,
  graph_problems_to = NULL, exclude_interval = NULL)
}
\arguments{
\item{path}{String giving a path to a parent directory containing species
and census datasets.}

\item{stem}{TRUE or FALSE, using the stem data (stem=TRUE) rather than the tree data (i.e. called 'full', stem=FALSE)}

\item{site}{the full name of your site (in lower case); e.g., 'barro colorado island'}

\item{dbh_units}{set the unit ("mm" or "cm") of DBH values, by default
dbh_units=="mm".}

\item{WD}{Optional, provide an external data.frame of wood densities by
species.}

\item{taper_correction}{TRUE or FALSE, apply Cushman et al (2014) taper correction}

\item{fill_missing}{TRUE or FALSE, interpolate missing DBH values}

\item{use_palm_allometry}{TRUE or FALSE, if TRUE, compute biomass of palm trees using a palm-specific allometric model from Goodman et al. (2013)}

\item{flag_stranglers}{TRUE or FALSE, individuals of known strangler fig species greater than 'dbh_stranglers' are flagged}

\item{dbh_stranglers}{(optional) minimal diameter (in mm) of strangler figs, default = 500}

\item{maxrel}{a numeric value: the threshold for flagging major errors in productivity, applied as absval(individual tree productivity)>maxrel*(average productivity per hectare)}

\item{write_errors_to}{A string giving a directory with the format
"path/to/file" (without extension) to output all records for trees with
major errors in productivity to a csv file. Defaults to not write such a
file.}

\item{graph_problems_to}{A string giving a directory with the format
"path/to/file" (without extension) to output graphs showing problematic
trees. The output may be multiple files, for example:
path/to/file_1.pdf,path/to/file_2.pdf, path/to/file_3.pdf, and so on.
Defaults to not write such files.}

\item{exclude_interval}{NULL by default. If needed a vector (e.g. c(1,2)) indicating which census interval(s) must be discarded from computation due, for instance, to a change in measurement protocol}

\item{output_errors}{TRUE or FALSE, output all records for trees with major errors in productivity to a csv file}

\item{DATA_path}{the pathname where the data are located}

\item{stem}{\code{TRUE}(default) or \code{FALSE} reflect that your censuses are
ForestGEO \code{stem} or \code{tree} (aka \code{full}) tables, respectively.}

\item{site}{String giving the name of your site -- one of one of
\code{site.info$site} (e.g., 'barro colorado island'.}

\item{taper_correction}{\code{TRUE} or \code{FALSE}, apply Cushman et al (2014) taper
correction.}

\item{fill_missing}{\code{TRUE} or \code{FALSE}, interpolate missing DBH values.}

\item{use_palm_allometry}{\code{TRUE} or \code{FALSE}, if \code{TRUE}, compute biomass of
palm trees using a palm-specific allometric model from Goodman et al.
(2013).}

\item{flag_stranglers}{\code{TRUE} or \code{FALSE}, individuals of known strangler fig
species greater than 'dbh_stranglers' are flagged}

\item{dbh_stranglers}{(optional) Minimal diameter (in mm) of strangler figs,
default = 500.}

\item{maxrel}{A numeric value: the threshold for flagging major errors in
productivity, applied as \code{absval(individual-tree-productivity) > maxrel * (average-productivity-per-hectare)}.}

\item{exclude_interval}{\code{NULL} by default. If needed a vector (e.g. c(1,2))
indicating which census interval(s) must be discarded from computation due,
for instance, to a change in measurement protocol.}
}
\value{
a data.table (data.frame) with all relevant variables.

A data.table (data.frame) with all relevant variables.
}
\description{
Main routine to format, detect major obvious errors, and gap-fill those errors in CTFS-formated data

Main routine to format, detect major obvious errors, and gap-fill those
errors in ForestGEO-like data.
}
\examples{
data_preparation(
  path = agb_example("data"),
  site = "barro colorado island",
  stem = TRUE
)
}
