% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{A concatenation of 3 mini "bci-stem_YEAR" datasets.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 645 rows and 38 columns.}
\usage{
df
}
\description{
\code{df} concatenates the 3 mini "bci-stem_YEAR" datasets. It avoids going
through lines.
}
\examples{
str(df)
}
\keyword{datasets}
