% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ficus}
\alias{ficus}
\title{Taxonomy for the genus ficus.}
\format{An object of class \code{data.frame} with 67 rows and 8 columns.}
\usage{
ficus
}
\description{
Taxonomy for the genus ficus.
}
\examples{
str(ficus)
}
\keyword{datasets}
