% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{flag_errors}
\alias{flag_errors}
\title{Flag major errors.}
\usage{
flag_errors(DF, site, flag_stranglers, maxrel, output_errors,
  exclude_interval)
}
\arguments{
\item{DF}{A data.table.}

\item{site}{Provide the full name of your site (in lower case) i.e. 'barro
colorado island'.}

\item{maxrel}{A numeric value setting the threshold over which relative
productivity is assumed to be too high (usually set at 20 percents).}

\item{output_errors}{\code{TRUE} or \code{FALSE}, output all records for trees with
major errors in productivity to a csv file.}

\item{exclude_interval}{A vector (i.e. c(1,2)) indicating if a set of census
intervals must be discarded from computation due for instance to a change
in  protocol of measurment.}

\item{strangler}{\code{TRUE} or \code{FALSE}, if \code{TRUE}  strangler figs tree are
flagged (upon a list to published soon).}
}
\value{
A data.table (data.frame) with all relevant variables.
}
\description{
Identify trees with major errors in DBH measurments. A major error correspond
to a relative individal productivity (or growth) is above a given percentage
(set by 'maxrel') of the mean productivity computed at a site. Additionnaly,
flagged trees that died at next census interval are also flagged. Option to
see DBH measurement (=draw.graph) of flagged trees or print a csv
(output_errors) are given.
}
