% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{format_interval}
\alias{format_interval}
\title{Format census intervals}
\usage{
format_interval(df, flag_stranglers, dbh_stranglers, code.broken = NULL)
}
\arguments{
\item{df}{a data.table}

\item{flag_stranglers}{TRUE or FALSE, individuals of known strangler fig species greater than 'dbh_stranglers' are flagged}

\item{dbh_stranglers}{(optional) minimal diameter (in mm) of strangler figs, default = 500}
}
\value{
a formated data.table.
}
\description{
Create census intervals (i.e. put consecutive census side by side), assign status by tree (i.e. alive (A),dead (D), recruited (R) or resprout (Rsp))
}
