% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_data}
\alias{list_data}
\title{Create a list from all .rda files in a parent directory.}
\usage{
list_data(parent)
}
\arguments{
\item{parent}{String giving the parent directory containing .rda files.}
}
\value{
A list where each element is the object stores in one .rda file
stored in \code{parent}.
}
\description{
Create a list from all .rda files in a parent directory.
}
\examples{
parent <- agb_example("data")
str(list_data(parent), list.len = 3)
}
