% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{loess.fun}
\alias{loess.fun}
\title{Loess.}
\usage{
loess.fun(x, var, range, weights = NULL)
}
\arguments{
\item{x}{A data.table.}

\item{var}{The name of the variable to be smoothed again intial AGB.}

\item{range}{The range of initial AGB to be used for prediction (i.e. 5th and
95th percentiles of the whole distribution).}
}
\value{
A smoothed prediction of the variable of interest.
}
\description{
A wrapper to get smoothed predictions of AGB fluxes using a loess function
(library 'locfit').
}
