% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sd_10}
\alias{sd_10}
\title{Dryad Wood Density Data standard deviation from BIOMASS pckg}
\format{An object of class \code{data.frame} with 3 rows and 2 columns.}
\usage{
sd_10
}
\description{
Dryad Wood Density Data standard deviation from BIOMASS pckg
}
\examples{
str(sd_10)
}
\keyword{datasets}
