% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site.info}
\alias{site.info}
\title{Information of ForestGEO sites.}
\format{An object of class \code{data.frame} with 63 rows and 12 columns.}
\usage{
site.info
}
\description{
Information of ForestGEO sites.
}
\examples{
str(site.info)
}
\keyword{datasets}
