% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdData}
\alias{wdData}
\title{Dryad Wood Density Data from BIOMASS pckg}
\format{An object of class \code{data.frame} with 16467 rows and 7 columns.}
\usage{
wdData
}
\description{
Dryad Wood Density Data from BIOMASS pckg
}
\examples{
str(wdData)
}
\keyword{datasets}
