% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WSG}
\alias{WSG}
\title{Density of woods at ForestGEO sites.}
\format{An object of class \code{data.frame} with 16558 rows and 9 columns.}
\usage{
WSG
}
\description{
Density of woods at ForestGEO sites.
}
\examples{
str(WSG)
}
\keyword{datasets}
