% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{computeAGB}
\alias{computeAGB}
\title{Biomass computation}
\usage{
computeAGB(df, site, DBH = NULL, use.CTFS.WD = T, H = NULL,
  use_palm_allometry)
}
\arguments{
\item{df}{the data.frame on which AGB shall be computed}

\item{DBH}{optional, specify the variable to be used (e.g. "dbh" or "dbh_corrected")}

\item{H}{optional, specify the column with height measurments}

\item{use_palm_allometry}{TRUE or FALSE, if TRUE, biomass of palm trees is computed through a specific allometric model (Goodman et al. 2013). Only valid for South America.}

\item{WD}{optional, provide an external data.frame of wood densities by species}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Allocate wood density and compute above-ground biomass using the updated model of Chave et al. (2014), given in Rejou-Mechain et al. (2017). Palm trees (palm=T) are computed using a different allometric model (Goodman et al. 2013).
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
