% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{determine_mean_prod}
\alias{determine_mean_prod}
\title{Set mean productivity.}
\usage{
determine_mean_prod(DF, site, flag_stranglers, exclude_interval)
}
\arguments{
\item{DF}{A data.table.}

\item{site}{Provide the full name of your site (in lower case) i.e. 'barro
colorado island'.}

\item{exclude_interval}{A vector (i.e. c(1,2)) indicating if a set of census
intervals must be discarded from computation due for instance to a change
in  protocol of measurment.}

\item{strangler}{\code{TRUE} or \code{FALSE}, if \code{TRUE}  strangler figs tree are
flagged (upon a list to published soon).}
}
\value{
Mean productivity in Mg/ha/yr.
}
\description{
Set mean productivity (Mg/ha/yr) across all census intervals for a given
site.
}
