% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{TODO: What is this dataset?}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 645 rows and 38 columns.}
\usage{
df
}
\description{
TODO: What is this dataset?
}
\examples{
str(df)
}
\keyword{datasets}
