% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{df}
\alias{df}
\title{A concatenation of 3 mini "bci-stem_YEAR" datasets.}
\usage{
df(x, df1, df2, ncp, log = FALSE)
}
\description{
\code{df} concatenates the 3 mini "bci-stem_YEAR" datasets. It avoids going
through lines.
}
\examples{
str(df)
}
