% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{LOAD}
\alias{LOAD}
\title{Load object}
\usage{
LOAD(saveFile)
}
\arguments{
\item{saveFile}{the path to the object to be loaded}
}
\value{
import the object
}
\description{
a wrapper to softly load R objects in the Global environment
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
